requirements.txt
anyio==4.9.0
asttokens==2.4.1
backcall==0.2.0
blis==0.7.10
Brotli==1.1.0
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
confection==0.1.4
cymem==2.0.8
debugpy==1.8.1
decorator==5.1.1
en-core-web-sm==3.8.0
exceptiongroup==1.2.0
executing==2.0.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.6
importlib_metadata==7.0.1
ipykernel==6.29.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.3
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
langcodes==3.4.0
language_data==1.2.0
language_tool_python==2.9.0
marisa-trie==1.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
murmurhash==1.0.10
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.0.2
packaging==24.0
parso==0.8.3
pickleshare==0.7.5
platformdirs==4.2.0
preshed==3.0.9
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.2
Pygments==2.17.2
PySocks==1.7.1
python-dateutil==2.9.0
python-dotenv==1.0.1
python-telegram-bot==22.0
pytz==2025.1
pywin32==308
pyzmq==25.1.2
regex==2024.11.6
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
smart_open==7.0.4
sniffio==1.3.1
spacy==3.8.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
thinc==8.2.2
toml==0.10.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.15.2
typer-slim==0.15.2
typing_extensions==4.11.0
ujson==5.10.0
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
win_inet_pton==1.1.0
